refactor: context 모듈화
